moved channel field into network.
